Finish all functionalities

+++ pages/api/products/[id].js
@@ -5,7 +5,10 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+
+  const token = cookies.token;
 
   await dbConnect();
 
@@ -19,7 +22,27 @@ export default async function handler(req, res) {
     }
   }
 
+  if (method === 'PUT') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json({ message: 'User not authorized!' });
+    }
+
+    try {
+      const product = await Product.findByIdAndUpdate(id, req.body, {
+        new: true,
+      });
+
+      res.status(200).json(product);
+    } catch (err) {
+      res.status(500).json(err);
+    }
+  }
+
   if (method === 'DELETE') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json({ message: 'User not authorized!' });
+    }
+
     try {
       await Product.findByIdAndDelete(id);
 
@@ -28,14 +51,4 @@ export default async function handler(req, res) {
       res.status(500).json(err);
     }
   }
-
-  // if (method === 'PUT') {
-  //   try {
-  //     const product = await Product.create(req.body);
-
-  //     res.status(201).json(product);
-  //   } catch (err) {
-  //     res.status(500).json(err);
-  //   }
-  // }
 }
